test: add button test for different modes
