+++ operate/client/src/App/ProcessInstance/BottomPanelTabs/VariablesTab/Variables/ViewFullVariableButton/Show/index.tsx
@@ -29,7 +29,7 @@ const ViewFullVariableButtonShow: React.FC<ViewFullVariableButtonShowProps> = ({
     <>
       <MaximizeButton
         label={`View full value of ${variableName}`}
-        onClick={async () => {
+        onClick={() => {
           setIsModalVisible(true);
         }}
       />
